+++ ClientApp/src/app/views/pages/login/login.component.ts
@@ -7,6 +7,9 @@ import { Component } from '@angular/core';
 })
 export class LoginComponent {
 
-  constructor() { }
+
+  constructor() {
+
+   }
 
 }
